==> 01-Data.Rmd <==
```{r data, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(gganimate)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
```

# Data

```{r datatypes, eval=F, echo=T}
load("data/Guyafor_Bota.RData")
vroom::vroom_write(Guyafor, "data/guyafor.csv", delim = ",")
csv2sql::csv_to_sqlite(csv_name = "data/guyafor.csv", db_name = "data/guyafor.sql", table_name = "inventory")
unlink("data/guyafor.csv")
```

```{r p6, eval=F}
guyafor <- DBI::dbConnect(RSQLite::SQLite(), dbname = "data/guyafor.sql")
p6 <- tbl(guyafor, "inventory") %>% 
  filter(Forest == "Paracou") %>% 
  filter(Plot == "6") %>% 
  collect()
DBI::dbDisconnect(guyafor)
g <- p6 %>% 
  ggplot(aes(Xfield, Yfield, size = CircCorr, col = paste(Genre, Espece))) +
  geom_point(aes(group = idTree), alpha = 0.5) +
  scale_color_discrete(guide = "none") +
  coord_equal() + 
  transition_states(CensusYear) + 
  ggtitle('Census Year : {closest_state}')
anim_save("save/p6.gif", g)
```

```{r p6gif, eval=F}
knitr::include_graphics("save/p6.gif")
```


```{r dataprep}
guyafor <- DBI::dbConnect(RSQLite::SQLite(), dbname = "data/guyafor.sql")
p6 <- tbl(guyafor, "inventory") %>% 
  filter(Forest == "Paracou") %>% 
  filter(Plot == "6") %>% 
  collect()
DBI::dbDisconnect(guyafor) ; rm(guyafor)
data <- p6 %>% 
  mutate(species = paste(Genre, Espece)) %>% 
   filter(!grepl("Indet", species)) %>% 
  mutate(speciesNum = as.numeric(as.factor(species))) %>% 
  mutate(DBH = CircCorr/pi) %>% 
  group_by(idTree) %>% 
  arrange(CensusYear) %>% 
  mutate(Y0 = first(CensusYear)) %>% 
  mutate(DBH0 = first(DBH)) %>% 
  mutate(Ytoday = last(CensusYear)) %>% 
  mutate(DBHtoday = last(DBH)+10^-6) %>% 
  group_by(Forest, Plot, SubPlot, TreeFieldNum, idTree, Xfield, Yfield, Xutm, Yutm, Family, Genus, Species,
           species, speciesNum, Y0, DBH0, Ytoday, DBHtoday) %>% 
  summarise(N = n()) %>% 
  filter(Ytoday == 2020) %>% 
  filter(DBHtoday > DBH0)
```

```{r datafig}
cowplot::plot_grid(
  data %>% 
    ggplot(aes(N)) +
    geom_histogram() +
    scale_y_log10() +
    geom_vline(xintercept = 10, col = "red") +
    xlab("Censuses per individual") +
    ggtitle(paste(round(nrow(filter(data, N > 10))/nrow(data)*100, 1), "% > 10")),
  data %>% 
    group_by(species) %>% 
    summarise(N = n()) %>% 
    ggplot(aes(N)) +
    geom_histogram() +
    scale_y_log10() +
    scale_x_log10() +
    geom_vline(xintercept = 10, col = "red") +
    xlab("Individuals per species") +
    ggtitle(paste(round(nrow(filter(summarise(group_by(data, species), N = n()), N > 10))/nrow(summarise(group_by(data, species), N = n()))*100, 1), "% > 10")),
  data %>% 
    ggplot(aes(Ytoday - Y0, DBHtoday - DBH0)) +
    geom_point() +
    xlab(expression(Delta[Y])) + ylab(expression(Delta[DBH]))
)
```

```{r mdata}
data <- data %>% 
  group_by(species) %>% 
  mutate(Nsp = n()) %>% 
  filter(N > 10, Nsp > 10) %>% 
  filter(species %in% sample(unique(data$species), 20)) %>% 
  ungroup() %>% 
  mutate(speciesNum = as.numeric(as.factor(species)))
mdata <- list(I = nrow(data),
              Y = 2020 - min(data$Y0) + 1,
              S = max(data$speciesNum),
              years = min(data$Y0):2020,
              DBH0 = data$DBH0,
              Y0 = data$Y0,
              DBHtoday = data$DBHtoday + 10^-6,
              species = data$speciesNum)
save(data, mdata, file = file.path("save", "mdata.Rdata")) 
```

```{r}
str(mdata)
```
