==> 01-ModelShape.Rmd <==
```{r model, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(cmdstanr)
library(rstan)
library(bayesplot)
library(loo)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F, eval=T)
```

# Model shape

In this chapter, I looked for the best model shape based on quality of fit, cross-validation (LOO), and prediction (RMSEP).
The result is summarised in Table \@ref(tab:mchoicetab).

## Data

I used reduced data to explore the model form. 
I focused on trees at 20 meters from any plot edges for neighbourhood effect.
I used only recruited trees in the censuses with at least 20 measurements of diameter at breast height (DBH, cm).
And, I randomly selected 20 trees for lightweight training data.
I further selected the last diameter measured for each individual as an evaluation data not used for model fitting (Fig. \@ref(fig:mmmdatafig)).

```{r data}
guyafor <- DBI::dbConnect(RSQLite::SQLite(), dbname = "data/guyafor.sql")
trees <- tbl(guyafor, "inventory") %>% 
  filter(Forest == "Paracou") %>% 
  collect() %>% 
  filter(Xfield > 20, Xfield < 250-20, Yfield > 20, Yfield < 250-20) %>% 
  mutate(species = paste(Genre, Espece)) %>% 
  filter(!grepl("Indet", species)) %>% 
  mutate(DBH = CircCorr/pi)
DBI::dbDisconnect(guyafor) ; rm(guyafor)
```

```{r mdata, eval=F}
trees %>% 
  group_by(Plot) %>% 
  mutate(StartYear = min(CensusYear)) %>% 
  group_by(idTree) %>% 
  arrange(CensusYear) %>% 
  filter(first(CensusYear) > StartYear) %>% 
  filter(first(DBH) < 12) %>% 
  filter(last(DBH) > first(DBH)) %>% 
  filter(n() > 20) %>% 
  mutate(Year = CensusYear - first(CensusYear)) %>%
  mutate(DBHprev = lag(DBH)) %>% 
  mutate(Deltat = CensusYear - lag(CensusYear)) %>% 
  mutate(DBHprev = ifelse(is.na(DBHprev), 10, DBHprev)) %>% 
  mutate(Deltat = ifelse(is.na(Deltat), 1, DBHprev)) %>% 
  ungroup() %>% 
  filter(idTree %in% sample(unique(.$idTree), 20)) %>%
  mutate(ind = as.numeric(as.factor(as.character(idTree)))) %>% 
  vroom::vroom_write("save/shape/trees_reduced.tsv")
vroom::vroom("save/shape/trees_reduced.tsv") %>% 
  group_by(idTree) %>% 
  filter(CensusYear == max(CensusYear)) %>% 
  vroom::vroom_write("save/shape/eval_reduced.tsv")
vroom::vroom("save/shape/trees_reduced.tsv") %>% 
  group_by(idTree) %>% 
  filter(CensusYear != max(CensusYear)) %>% 
  vroom::vroom_write("save/shape/mdata_reduced.tsv")
```

```{r mmmdatafig, fig.cap="Tree diameter trajectories in training data. Color represent individuals and the red point the data used for evaluation."}
ggplot(vroom::vroom("save/shape/mdata_reduced.tsv"),
       aes(Year, DBH, col = as.factor(ind))) +
  geom_point() +
  geom_point(data = vroom::vroom("save/shape/eval_reduced.tsv"), col = "red", size = 2) +
  geom_smooth(se = F) +
  scale_color_discrete(guide = "none") +
  xlab("Years since recruitment") + ylab("Diameter at breast hieght (cm)")
```

## Michaelis Menten

The first tested model is a Michaelis Menten model using time $t$ since recruitment (DBH=10cm) for each individual tree $i$:

$$DBH_{i,t} \sim \mathcal N (10 + \frac{\alpha_i \times t}{\beta_i+t}, \sigma)$$
where $\alpha$ is individual maximum diameter and $\beta$ the year since recruitment when the individual reaches half the maximum diameter value.
The model correctly converged ($\hat R < 1.05$) with an acceptable but marked correlation between $alpha$ and $beta$.
$beta$ posteriors have not too much uncertainty and are varying widely among individuals.
Finally predictions of the diameter trajectories by the model are good.

```{r mmstanfit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
mm <- cmdstan_model("model/mm.stan")
fit <- mm$sample(
                  data = list(
                  N = nrow(mdata),
                  I = max(mdata$ind),
                  Y = max(mdata$Year)+1,
                  year = mdata$Year+1,
                  ind = mdata$ind,
                  dbh = mdata$DBH,
                  Np = nrow(eval),
                  dbhp = eval$DBH,
                  indp = eval$ind,
                  yearp = eval$Year),
  chains = 2, 
  parallel_chains = 2,
  refresh = 200,
  save_warmup = F
)
fit_mm <- rstan::read_stan_csv(fit$output_files())
save(fit_mm, file = "save/shape/mm.Rdata")
```

```{r mmsampling}
load("save/shape/mm.Rdata")  
```

```{r mmrhat}
mcmc_rhat(rhat(fit_mm)) 
```

```{r mmtrace}
mcmc_trace(fit_mm, pars = c("alpha[1]", "beta[1]", "sigma")) 
```

```{r mmpairs}
mcmc_pairs(fit_mm, pars = c("alpha[1]", "beta[1]", "sigma")) 
```

```{r mmbeta}
mcmc_intervals(fit_mm, regex_pars = "beta") 
```

```{r mmpreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
pars <- as.data.frame(fit_mm, pars = c("alpha", "beta")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
preds <- pars %>% 
  mutate(Year = list(seq(0, max(mdata$Year), length.out = 100))) %>% 
  unnest(Year) %>% 
  mutate(DBH = 10 + (alpha * Year) / (beta + Year))
ggplot(left_join(mdata, pars), aes(Year, DBH, col = log(beta), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(data = preds) +
  viridis::scale_color_viridis() 
```

## Gompertz

The second tested model uses a Gompertz model [@Herault2011], were the diameter of individual $i$ at year $t$ is the individual diameter from previous year plus its annual growth:

$$DBH_{i,s,t} \sim \mathcal N (DBH_{i,t-i} + Gmax_i \times  exp(-\frac12.[\frac{log(\frac{DBH_{i,t-i}}{100.Dopt_i})}{Ks_i}]^2)\times\Delta t, \sigma)$$

The annual growth rate for individual $i$ at year $y$ with a diameter of $DBH_{y,i}$ is defined following a Gompertz model [@Gompertz1825] already identified as the best model for growth-trajectories in Paracou [@Herault2011],
where $Gmax$ is the maximum growth potential,
$Dopt$ is the optimal diameter at which the individual reaches its maximum growth potential,
and $Ks$ is the kurtosis defining the width of the bell-shaped growth-trajectory [see figure 1 in @Herault2011].
On model parameter diverged ($\hat R > 1.05$) with no correlations among parameters.
$gmax$ posteriors are highly uncertain and does not vary widely among individuals.
Finally predictions of the diameter trajectories by the model are overfitting because the model uses the diameter from previous census reporting the measurement error from one year to the other.

```{r gompertzstanfit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
gompertz <- cmdstan_model("model/gompertz.stan")
fit <- gompertz$sample(
  data = list(
    N = nrow(mdata),
    I = max(mdata$ind),
    dbhprev = mdata$DBHprev,
    dt = mdata$Deltat,
    ind = mdata$ind,
    dbh = mdata$DBH,
    Np = nrow(eval),
    dbhp = eval$DBH,
    indp = eval$ind,
    dbhprevp = eval$DBHprev,
    dtp = eval$Deltat),
  chains = 2, 
  parallel_chains = 2,
  refresh = 200,
  save_warmup = F
)
fit_gompertz <- rstan::read_stan_csv(fit$output_files())
save(fit_gompertz, file = "save/shape/gompertz.Rdata")
```

```{r gompertzsampling}
load("save/shape/gompertz.Rdata")  
```

```{r gompertzrhat}
mcmc_rhat(rhat(fit_gompertz)) 
```

```{r gompertztrace}
mcmc_trace(fit_gompertz, pars = c("gmax[1]", "dopt[1]", "ks[1]", "sigma")) 
```

```{r gompertzpairs}
mcmc_pairs(fit_gompertz, pars = c("gmax[1]", "dopt[1]", "ks[1]", "sigma")) 
```

```{r gompertzgmax}
mcmc_intervals(fit_gompertz, regex_pars = "gmax") 
```

```{r gompertzpreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
pars <- as.data.frame(fit_gompertz, pars = c("gmax", "dopt", "ks")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
left_join(mdata, pars) %>% 
  mutate(DBHpred = DBHprev + gmax * exp(-0.5* (log(DBHprev / (100*dopt)) / ks)^2) * Deltat) %>% 
  ggplot(aes(Year, DBH, col = log(gmax), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(aes(y = DBHpred)) +
  viridis::scale_color_viridis() 
```

## Lognormal

The third tested model is a lognormal model using time $t$ since recruitment (DBH=10cm) for each individual tree $i$:

$$DBH_{i,t} \sim \mathcal N (10 +\beta \times log(t), \sigma)$$
where $\beta$ is the slope of year since recruitment effect on the log response of diameter to time.
The model correctly converged ($\hat R < 1.05$) with no correlations among parameters.
$beta$ posteriors have not too much uncertainty and are varying widely among individuals.
Finally predictions of the diameter trajectories by the model are a bit far from observations.

```{r logfit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
log <- cmdstan_model("model/lognormal.stan")
fit <- log$sample(
  data = list(
    N = nrow(mdata),
    I = max(mdata$ind),
    Y = max(mdata$Year)+1,
    year = mdata$Year+1,
    ind = mdata$ind,
    dbh = mdata$DBH,
    Np = nrow(eval),
    dbhp = eval$DBH,
    indp = eval$ind,
    yearp = eval$Year),
  chains = 2, 
  parallel_chains = 2,
  refresh = 200,
  save_warmup = F
)
fit_log <- rstan::read_stan_csv(fit$output_files())
save(fit_log, file = "save/shape/log.Rdata")
```

```{r logsampling}
load("save/shape/log.Rdata")  
```

```{r logrhat}
mcmc_rhat(rhat(fit_log)) 
```

```{r logtrace}
mcmc_trace(fit_log, pars = c("beta[1]", "sigma")) 
```

```{r logpairs}
mcmc_pairs(fit_log, pars = c("beta[1]", "sigma")) 
```

```{r loggmax}
mcmc_intervals(fit_log, regex_pars = "beta") 
```

```{r logpreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
pars <- as.data.frame(fit_log, pars = c("beta")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
preds <- pars %>% 
  mutate(Year = list(seq(0, max(mdata$Year), length.out = 100))) %>% 
  unnest(Year) %>% 
  mutate(DBH = 10 + beta * log(Year+1))
ggplot(left_join(mdata, pars),
       aes(Year, DBH, col = log(beta), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(data = preds) +
  viridis::scale_color_viridis() 
```

## Polynomial

The fourth tested model is a third degree polynomial model using time $t$ since recruitment (DBH=10cm) for each individual tree $i$:

$$DBH_{i,t} \sim \mathcal N (\alpha \times t^3 + \beta_i \times t^2+ \gamma_i \times t + 10, \sigma)$$

where $\alpha$, $\beta$, and $\gamma$ are the slope of year since recruitment effects to its cube, square or identity.
The model correctly converged ($\hat R < 1.05$) with an acceptable but marked correlation between $\alpha$, $\beta$, and $\gamma$.
$gamma$ posteriors have not too much uncertainty and are varying widely among individuals.
Finally predictions of the diameter trajectories by the model are a bit overfitted with unrealistic inflections at the end of the trajectories


```{r polyfit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
poly <- cmdstan_model("model/poly3.stan")
fit <- poly$sample(
  data = list(
    N = nrow(mdata),
    I = max(mdata$ind),
    Y = max(mdata$Year)+1,
    year = mdata$Year+1,
    ind = mdata$ind,
    dbh = mdata$DBH,
    Np = nrow(eval),
    dbhp = eval$DBH,
    indp = eval$ind,
    yearp = eval$Year),
  chains = 2, 
  parallel_chains = 2,
  refresh = 200,
  save_warmup = F
)
fit_poly <- rstan::read_stan_csv(fit$output_files())
save(fit_poly, file = "save/shape/poly.Rdata") 
```

```{r polysampling}
load("save/shape/poly.Rdata")  
```

```{r polyrhat}
mcmc_rhat(rhat(fit_poly)) 
```

```{r polytrace}
mcmc_trace(fit_poly, pars = c("alpha[1]",  "beta[1]", "gamma[1]", "sigma")) 
```

```{r polypairs}
mcmc_pairs(fit_poly, pars = c("alpha[1]",  "beta[1]", "gamma[1]", "sigma")) 
```

```{r polyalpha}
mcmc_intervals(fit_poly, regex_pars = "gamma") 
```

```{r polypreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
pars <- as.data.frame(fit_poly, pars = c('alpha', "beta", "gamma")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
preds <- pars %>% 
  mutate(Year = list(seq(0, max(eval$Year), length.out = 100))) %>% 
  unnest(Year) %>% 
  mutate(DBH = alpha * Year^3 + beta * Year^2 + gamma * Year + 10)
ggplot(left_join(mdata, pars), 
       aes(Year, DBH, col = log(gamma), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(data = preds) +
  viridis::scale_color_viridis() 
```

## Weibul

The fifth tested model is a Weibul model using time $t$ since recruitment (DBH=10cm) for each individual tree $i$:

$$DBH_{i,t} \sim \mathcal N (10 + \alpha \times (1 − e^{-t_i/\beta}) , \sigma)$$
The model correctly converged ($\hat R < 1.05$) with an acceptable but marked correlation between $\alpha$ and $\beta$.
$beta$ posteriors have not too much uncertainty and are varying widely among individuals.
Finally predictions of the diameter trajectories by the model are good.

```{r weibulfit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
weibul <- cmdstan_model("model/weibul.stan")
fit <- weibul$sample(
  data = list(
    N = nrow(mdata),
    I = max(mdata$ind),
    Y = max(mdata$Year)+1,
    year = mdata$Year+1,
    ind = mdata$ind,
    dbh = mdata$DBH,
    Np = nrow(eval),
    dbhp = eval$DBH,
    indp = eval$ind,
    yearp = eval$Year),
  chains = 2, 
  parallel_chains = 2,
  refresh = 200,
  save_warmup = F
)
fit_weibul <- rstan::read_stan_csv(fit$output_files())
save(fit_weibul, file = "save/shape/weibul.Rdata") 
```

```{r weibulsampling}
load("save/shape/weibul.Rdata")  
```

```{r weibulrhat}
mcmc_rhat(rhat(fit_weibul)) 
```

```{r weibultrace}
mcmc_trace(as.array(fit_weibul, pars = c("alpha[1]", "beta[1]", "sigma")))
```

```{r weibulpairs}
mcmc_pairs(as.array(fit_weibul, pars = c("alpha[1]", "beta[1]", "sigma"))) 
```

```{r weibulbeta}
mcmc_intervals(as.array(fit_weibul, pars = "beta"))
```

```{r weibulpreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
pars <- as.data.frame(fit_weibul, pars = c('alpha', "beta")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
preds <- pars %>% 
  mutate(Year = list(seq(0, max(eval$Year), length.out = 100))) %>% 
  unnest(Year) %>% 
  mutate(DBH = 10 + alpha * (1 - exp(-(Year+1)/beta) ))
ggplot(left_join(mdata, pars), 
       aes(Year, DBH, col = log(beta), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(data = preds) +
  viridis::scale_color_viridis() 
```

## Amani

The sixth tested model was developed by @Amani2021 and uses time $t$ since recruitment (DBH=10cm) for each individual tree $i$:

$$DBH_{i,t} \sim \mathcal N (\alpha \times (1 − e^{-\lambda \times (\frac{t_i}{\theta})^\beta}) , \sigma)$$
The model correctly converged ($\hat R < 1.05$) with an acceptable but marked correlation between $\alpha$, $\beta$, and $\lambda$.
$\theta$ posteriors are highly uncertain and does not vary widely among individuals.
Finally predictions of the diameter trajectories by the model are good but the inflection point if only used in the yellow trajectory more probably due to a change of POM than real growth change.

```{r amanifit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
amani <- cmdstan_model("model/amani.stan")
fit <- amani$sample(
  data = list(
    N = nrow(mdata),
    I = max(mdata$ind),
    Y = max(mdata$Year)+1,
    year = mdata$Year+1,
    ind = mdata$ind,
    dbh = mdata$DBH,
    Np = nrow(eval),
    dbhp = eval$DBH,
    indp = eval$ind,
    yearp = eval$Year),
  chains = 2, 
  parallel_chains = 2,
  refresh = 200,
  save_warmup = F
)
fit_amani <- rstan::read_stan_csv(fit$output_files())
save(fit_amani, file = "save/shape/amani.Rdata") 
```

```{r amanisampling}
load("save/shape/amani.Rdata")  
```

```{r amanirhat}
mcmc_rhat(rhat(fit_amani)) 
```

```{r amanitrace}
mcmc_trace(as.array(fit_amani, pars = c("alpha[1]", "beta[1]", "lambda[1]", "theta[1]", "sigma")))
```

```{r amanipairs}
mcmc_pairs(as.array(fit_amani, pars = c("alpha[1]", "beta[1]", "lambda[1]", "theta[1]", "sigma"))) 
```

```{r amanibeta}
mcmc_intervals(as.array(fit_amani, pars = "theta"))
```

```{r  amanpreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
pars <- as.data.frame(fit_amani, pars = c('alpha', "beta", "lambda", "theta")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
preds <- pars %>% 
  mutate(Year = list(seq(0, max(eval$Year), length.out = 100))) %>% 
  unnest(Year) %>% 
  mutate(DBH = 10 + alpha * (1 - exp(-lambda*((Year+1)/ theta)^beta )))
ggplot(left_join(mdata, pars), 
       aes(Year, DBH, col = log(beta), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(data = preds) +
  viridis::scale_color_viridis() 
```

## Gompertz sum

The seventh tested model uses a Gompertz model [@Herault2011], were the diameter of individual $i$ at year $t$ is the sum of annual growth from $t0$ to $t$:

$$DBH_{t,i} \sim \mathcal N  (10 + gmax_i \times \sum _{y=1|DBH_{t=0}} ^{y=t} exp(-\frac12.[\frac{log(\frac{DBH_{t,i}}{100.Dopt_i})}{Ks_i}]^2)), \sigma)$$

The annual growth rate for individual $i$ at year $y$ with a diameter of $DBH_{y,i}$ is defined following a Gompertz model [@Gompertz1825] already identified as the best model for growth-trajectories in Paracou [@Herault2011],
where $Gmax$ is the maximum growth potential,
$Dopt$ is the optimal diameter at which the individual reaches its maximum growth potential,
and $Ks$ is the kurtosis defining the width of the bell-shaped growth-trajectory [see figure 1 in @Herault2011].
The model correctly converged ($\hat R < 1.05$) with an acceptable but marked correlation between $dopt$ and $ks$.
$gmax$ posteriors have logical uncertainty but are varying widely among individuals.
Finally predictions of the diameter trajectories by the model are good and realistic.
Moreover, model errors do not show temporal autocorrelation per individual.

```{r gompertzsumfit, eval=F}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
gompertzsum <- cmdstan_model("model/gompertzsum.stan")
fit <- gompertzsum$sample(
  data = list(
    N = nrow(mdata),
    I = max(mdata$ind),
    Y = max(max(mdata$Year), max(eval$Year))+1,
    year = mdata$Year+1,
    ind = mdata$ind,
    dbh = mdata$DBH,
    Np = nrow(eval),
    dbhp = eval$DBH,
    indp = eval$ind,
    yearp = eval$Year+1),
  chains = 2, 
  parallel_chains = 4,
  refresh = 200,
  save_warmup = F
)
fit_gompertzsum <- rstan::read_stan_csv(fit$output_files())
save(fit_gompertzsum, file = "save/shape/gompertzsum.Rdata") 
```

```{r gompertzsumsampling}
load("save/shape/gompertzsum.Rdata")  
```

```{r gompertzsumrhat}
mcmc_rhat(rhat(fit_gompertzsum)) 
```

```{r gompertzsumtrace}
mcmc_trace(as.array(fit_gompertzsum, pars = c("gmax[1]", "dopt[1]", "ks[1]", "sigma")))
```

```{r gompertzsumpairs}
mcmc_pairs(as.array(fit_gompertzsum, pars = c("gmax[1]", "dopt[1]", "ks[1]", "sigma"))) 
```

```{r gompertzsumalpha}
mcmc_intervals(as.array(fit_gompertzsum, pars = "gmax"))
```

```{r gompertzsumpreds}
mdata <- vroom::vroom("save/shape/mdata_reduced.tsv")
eval <- vroom::vroom("save/shape/eval_reduced.tsv")
dbhtoday <- function(t, gmax, dopt, ks){
  dbh <- 10
  for(i in 1:t)
    dbh <- dbh + gmax*exp(-0.5* (log(dbh / (100*dopt)) / ks)^2)
  return(dbh)  
}
pars <- as.data.frame(fit_gompertzsum, pars = c('gmax', "dopt", "ks")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
preds <- pars %>% 
  mutate(Year = list(0:max(eval$Year))) %>% 
  unnest(Year) %>% 
  rowwise() %>% 
  mutate(DBH = dbhtoday(Year+1, gmax, dopt, ks))
ggplot(left_join(mdata, pars), aes(Year, DBH, col = log(gmax), group = ind)) +
  geom_point() +
  geom_point(data = eval, col = "red", size = 2) +
  geom_line(data = preds) +
  viridis::scale_color_viridis() 
```

```{r gompertzsumres}
dbhtoday <- function(t, gmax, dopt, ks){
  dbh <- 10
  for(i in 1:t)
    dbh <- dbh + gmax*exp(-0.5* (log(dbh / (100*dopt)) / ks)^2)
  return(dbh)  
}
pars <- as.data.frame(fit_gompertzsum, pars = c('gmax', "dopt", "ks")) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
left_join(vroom::vroom("save/shape/mdata_reduced.tsv"), pars) %>% 
  rowwise() %>% 
  mutate(DBHpred = dbhtoday(Year+1, gmax, dopt, ks)) %>% 
  mutate(Error = DBH - DBHpred) %>% 
  ggplot(aes(Year, Error, col = as.character(ind))) +
  geom_smooth(se = F, alpha = 0.1) +
  geom_point() +
  scale_color_discrete(guide = "none")
```

## Comparisons

The Gompertz sum model has the best prediction (lowest RMSEP) associated to the second best evaluation (second lowest loo epld), knowing that the first one use every year diameter and overfit, and a decent computing time. This is the best model that I'll use in next steps.

```{r fits}
fits <- list(
  "Michaelis Menten" = fit_mm,
  "Gompertz" = fit_gompertz,
  "Lognormal" = fit_log,
  "Polynomial" = fit_poly,
  "Weibul" = fit_weibul,
  "Amani" = fit_amani,
  "Gompertz sum" = fit_gompertzsum
)
```

```{r mchoicetab}
lapply(fits, mcmc_intervals_data, "RMSEP") %>% 
  bind_rows(.id = "Model") %>% 
  dplyr::rename(RMSEP = m) %>% 
  dplyr::select(Model, RMSEP) %>% 
  left_join(lapply(fits, loo) %>% 
              loo_compare() %>% 
              as.data.frame() %>% 
              rownames_to_column("Model") %>% 
              dplyr::select(Model, elpd_loo)) %>% 
  left_join(lapply(fits, get_elapsed_time) %>% 
              lapply(as.data.frame) %>% 
              bind_rows(.id = "Model") %>% 
              rowwise() %>% 
              mutate(Time = sum(warmup + sample)) %>% 
              dplyr::select(Model, Time) %>% 
              group_by(Model) %>% 
              summarise(Time = max(Time))) %>% 
  kable(caption = "Model choice summary with quality of prediciton (Root Mean Square Error of Prediction, RMSEP), cross-validation (Leave-One-Out Estimate of the expected Log pointwise Predictive Density, LOO ELDP), and speed (Elapsed yime).",
        col.names = c("Model", "RMSEP", "LOO ELDP", "Elapsed time"))
```

```{r rmsep, fig.cap="Models root mean square error of predicition. The X scale is sqaure-root transformed."}
lapply(fits, mcmc_intervals_data, "RMSEP") %>% 
  bind_rows(.id = "Model") %>% 
  ggplot(aes(x = Model, xend = Model)) +
  geom_point(aes(y = m), shape = 21, size = 3, alpha = 0.5) +
  geom_segment(aes(y = ll, yend = hh),
               size = 1, show.legend = F, alpha = 0.5) +
  geom_segment(aes(y = l, yend = h), size = 2, alpha = 0.5) +
  coord_flip() + xlab("") + scale_y_log10() + ylab(expression(sqrt(bar((Y-hat(Y))^2))))
```

```{r loo}
lapply(fits, loo) %>% 
  loo_compare() %>%
  kable(caption = "Leave one out evaluation of models (http://cran.nexr.com/web/packages/loo/vignettes/loo-example.html).")
```

```{r time, fig.cap="Model elapsed time."}
lapply(fits, get_elapsed_time) %>% 
  lapply(as.data.frame) %>% 
  bind_rows(.id = "Model") %>% 
  rowwise() %>% 
  mutate(Time = sum(warmup + sample)) %>% 
  group_by(Model) %>% 
  summarise(Time = max(Time)) %>% 
  ggplot(aes(Model, Time)) +
  geom_point() +
  coord_flip() +
  scale_y_log10() +
  ylab("Time (seconds)")
```
