==> 02a-Growth.Rmd <==
```{r setupgrowtha, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(bayesplot)
library(rstan)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F, eval=T)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = T)
```

# Growth `stan`

For the first model we used a lognormal likelihood that we sampled with the NUTS algorithm using `stan`. 
Specifically, we used the following model to estimate individual growth potential:

\begin{equation} 
  DBH_{y=today,s,i}  - DBH_{y=y0,s,i} \sim \\
  \mathcal{logN} (log(\theta_{1,s} \times \sum _{y=y_0} ^{y=today} exp(-\frac12.[\frac{log(\frac{DBH_{y,s,i}}{100.\theta_{2,s}})}{\theta_{3,s}}]^2)), \sigma_1) \\ 
  ~ \\
  (\theta_{1,s}, \theta_{2,s}, \theta_{3,s}) \sim \mathcal{logN}^{3 \times S}(log(1),1) \\ 
  \sigma \sim \mathcal{N}_T(0,1) \\ 
  ~ \\
  \theta_{1,s,i} = \frac{DBH_{y=today,s,i}}{\sum_{y=y0}^{y=2020} exp(-\frac12.[\frac{log(\frac{DBH_{y,s,i}}{100.\theta_{2,s}})}{\theta_{3,s}}]^2)} \\ 
    (\#eq:gmax1)
\end{equation} 

The sampling took up to 46 minutes (Tab. \@ref(tab:timea)) with chains correctly mixing (Fig. \@ref(fig:traceplota)) and relatively no correlations among parameters (Fig. \@ref(fig:pairsa)). 
We obtained correct and varying species (Fig. \@ref(fig:thetasa)) and individual (Fig. \@ref(fig:thetaia)) growth potentials.

```{r samplinga, eval=F}
load("save/mdata_reduced.Rdata")
model <- stan_model("model/gmax.stan")
fit <- sampling(model, chains = 2, data = mdata, save_warmup = F, control = list(adapt_delta = 0.99, max_treedepth = 12),
                include = F, pars = c("epsilon_s2", "epsilon_s3", "epsilon_i", "DBH"))
save(data, mdata, fit, file = "save/gmax_stan_reduced.Rdata"))
```

```{r fita}
load("save/gmax_stan_reduced.Rdata") 
```

```{r timea}
get_elapsed_time(fit) %>% 
  as.data.frame() %>% 
  rownames_to_column("chain") %>% 
  mutate(total = warmup + sample) %>% 
  mutate_at(c("warmup", "sample", "total"), lubridate::seconds_to_period) %>% 
  kable(caption = "Elapsed time to fit 187 trees over 37 years and 3 species in the plot 14 of Paracou.")
```

```{r traceplota, fig.cap="Trace plot."}
mcmc_trace(fit, pars = c("thetas[1,1]", "thetas[1,2]", "thetas[1,3]",  "sigma", "lp__"), np = nuts_params(fit))
```

```{r pairsa, fig.cap="Pairs plot."}
mcmc_pairs(fit, pars = c("thetas[1,1]", "thetas[1,2]", "thetas[1,3]",  "sigma"), np = nuts_params(fit))
```

```{r thetasa, fig.cap="Posteriors for species growth potentials."}
mcmc_intervals(fit, pars = paste0("thetas[", 1:mdata$S, ",1]"))
```

```{r thetaia, fig.cap="Posteriors for individual growth potentials."}
mcmc_intervals(fit, pars = paste0("thetai[", 1:mdata$I, "]")) +
  theme(axis.text.y = element_blank())
```
