==> 02b-Growth.Rmd <==
```{r setupgrowthc, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(bayesplot)
library(greta)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F, eval=T)
```

# Growth `greta`

For the second model we used a lognormal likelihood that we sampled with the HMC  algorithm and variational inference using `greta`. 
Specifically, we used the following model to estimate individual growth potential:

\begin{equation} 
  DBH_{y=today,s,i}  - DBH_{y=y0,s,i} \sim \\
  \mathcal{logN} (log(\theta_{1,s} \times \sum _{y=y_0} ^{y=today} exp(-\frac12.[\frac{log(\frac{DBH_{y,s,i}}{100.\theta_{2,s}})}{\theta_{3,s}}]^2)), \sigma_1) \\ 
  ~ \\
  (\theta_{1,s}, \theta_{2,s}, \theta_{3,s}) \sim \mathcal{logN}^{3 \times S}(log(1),1) \\ 
  \sigma \sim \mathcal{N}_T(0,1) \\ 
  ~ \\
  \theta_{1,s,i} = \frac{DBH_{y=today,s,i}}{\sum_{y=y0}^{y=2020} exp(-\frac12.[\frac{log(\frac{DBH_{y,s,i}}{100.\theta_{2,s}})}{\theta_{3,s}}]^2)} \\ 
    (\#eq:gmax2)
\end{equation} 

The sampling took only 15 minutes (Tab. \@ref(tab:timec)) with chains correctly mixing but encompassing static state (Fig. \@ref(fig:traceplotc)) and relatively no correlations among parameters (Fig. \@ref(fig:pairsc)). 
We obtained correct and varying species (Fig. \@ref(fig:thetasc)) and individual (Fig. \@ref(fig:thetaic)) growth potentials.

```{r samplingc, eval=F}
load("save/mdata_reduced.Rdata")

# variables & priors
thetas <- lognormal(0, 1, dim = c(mdata$S, 3), truncation = c(0, 1))
sigma <- normal(0, 1, truncation = c(0, Inf))

# predictor
Sigma <- rep(1, mdata$I)
for(t in 1:mdata$Y) {
  Sigma[mdata$Y0 == mdata$years[t]] <- mdata$DBH0[mdata$Y0 == mdata$years[t]]
  Sigma <- Sigma + exp(-0.5* (log(Sigma / (100*thetas[mdata$species,2])) / thetas[mdata$species,3])^2)
}
Sigma <- Sigma - mdata$DBH0

# likelihood
DeltaDBH <- mdata$DBHtoday - mdata$DBH0
distribution(DeltaDBH) <- lognormal(log(thetas[mdata$species,1]*Sigma), sigma)

# generate quantities
thetai <- mdata$DBHtoday / Sigma

# model
m <- model(thetai, thetas, sigma)

# sampling
fit <- mcmc(m, warmup = 1000, n_samples = 1000, chains = 2)

save(data, mdata, fit, file = "save/gmax_greta_reduced.Rdata")
```

```{r fitc}
load("save/gmax_greta_reduced.Rdata") 
```

```{r timec}
data.frame("Total" = "< 20 minutes") %>% 
  kable(caption = "Elapsed time to fit 187 trees over 37 years and 3 species in the plot 14 of Paracou.") 
```

```{r traceplotc, fig.cap="Trace plot."}
mcmc_trace(fit, pars = c("thetas[1,1]", "thetas[1,2]", "thetas[1,3]", "sigma")) 
```

```{r pairsc, fig.cap="Pairs plot."}
mcmc_pairs(fit, pars = c("thetas[1,1]", "thetas[1,2]", "thetas[1,3]", "sigma")) 
```

```{r thetasc, fig.cap="Posteriors for species growth potentials."}
mcmc_intervals(fit, pars = paste0("thetas[", 1:mdata$S, ",1]")) 
```

```{r thetaic, fig.cap="Posteriors for individual growth potentials."}
mcmc_intervals(fit, pars = paste0("thetai[", 1:mdata$I, ",1]")) +
  theme(axis.text.y = element_blank()) 
```
