==> 03-ModelChoice.Rmd <==
```{r setupmodelchoice, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(bayesplot)
library(rstan)
library(greta)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F, eval=T)
```

# Model choice

We fitted two models on reduced data:

* **stan lognormal**: a lognormal likelihood that we sampled with the NUTS algorithm using `stan`
* **greta lognormal**: a lognormal likelihood that we sampled with the NUTS algorithm using `greta`

`greta` is quicker for sampling (Tab. \@ref(tab:timesreduced)) but its chains behave less well than `stan` (Fig. \@ref(fig:traceplotsreduced)).
And growth potentials were underestimated when using `greta` compared to `stan` (Fig. \@ref(fig:thetasreduced) and Fig. \@ref(fig:thetaireduced)).
Consequently, we will use the stan lognormal model for further analyses which is the second fastest and which uses the most appropriate likelihood.

```{r fitsreduced}
load("save/gmax_stan_reduced.Rdata")
fit1 <- fit
load("save/gmax_greta_reduced.Rdata")
fit2 <- fit
rm(fit)
```

```{r traceplotsreduced, fig.cap="Trace plot across models."}
lapply(list("stan" = fit1, "greta" = fit2), mcmc_trace_data, pars = c("thetas[1,1]", "thetas[1,2]", "thetas[1,3]", "sigma")) %>%
  bind_rows(.id = "model") %>% 
  ggplot(aes(iteration, value, col = model)) +
  geom_line(alpha = 0.5) +
  facet_grid(parameter ~ chain, scales = "free")
```

```{r thetasreduced, fig.cap="Posteriors for species growth potentials across models."}
lapply(list("stan" = fit1, "greta" = fit2), mcmc_intervals_data, pars = paste0("thetas[", 1:mdata$S, ",1]")) %>%
  bind_rows(.id = "model") %>% 
  ggplot(aes(x = parameter, 
             xend = parameter,
             color = model, fill = model)) +
  geom_jitter(aes(y = m), shape = 21, size = 3, alpha = 0.5) +
  geom_segment(aes(y = ll, yend = hh),
               size = 1, show.legend = F, alpha = 0.5) +
  geom_segment(aes(y = l, yend = h), size = 2, alpha = 0.5) +
  coord_flip() +
  xlab(expression(theta[s])) + ylab("Posterior")
```
