==> 05b-Analyses.Rmd <==
```{r analysesb, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(bayesplot)
library(ggtree)
library(phylosignal)
library(adephylo)
library(ape)
library(phylobase)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F, eval=T)
```

```{r envb, eval=F}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(parallel)
library(doSNOW)
library(foreach)
library(tidyverse)
data <- vroom::vroom("save/mdatafull.tsv") %>% 
  dplyr::select(idTree, Plot, Xutm, Yutm) %>% 
  unique()
cl <- makeCluster(4, outfile = "/dev/null")
registerDoSNOW(cl)
pb <- txtProgressBar(max = nrow(data), style = 3)
progress <- function(n) setTxtProgressBar(pb, n)
opts <- list(progress = progress)
NC <- foreach(ind=1:nrow(data),
              .packages = "dplyr",
              .options.snow = opts) %dopar% {
                con <- DBI::dbConnect(RSQLite::SQLite(), dbname = "data/guyafor.sql")
                NC <- tbl(con, "inventory") %>% 
                  filter(Plot == local(data$Plot[ind])) %>% 
                  filter(idTree != local(data$idTree[ind])) %>% 
                  mutate(dij = sqrt((local(data$Xutm[ind]) - Xutm)^2+(local(data$Yutm[ind]) - Yutm)^2)) %>% 
                  filter(dij < 20) %>% 
                  mutate(DBH = CircCorr/pi) %>% 
                  collect() %>% 
                  group_by(CensusYear) %>% 
                  summarise(NCI = sum(DBH*DBH*exp(-0.25*dij))) %>% 
                  ungroup() %>% 
                  summarise(idTree = local(data$idTree[ind]),
                            NCI = mean(NCI))
                DBI::dbDisconnect(con)
                return(NC)
              } 
close(pb)
stopCluster(cl)
NC <- bind_rows(NC)
XY <- NC %>% 
  left_join(data) %>% 
  sf::st_as_sf(coords = c("Xutm", "Yutm"),
             crs = '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0')
NC$twi <- raster::extract(raster::raster("data/TWI_1m.tif"), XY)
write_tsv(NC, file = "save/env_full.tsv")
```

```{r phylob, eval=F}
library(V.PhyloMaker)
splist <- vroom::vroom("save/mdatafull.tsv") %>% 
  dplyr::select(Family, Genus, species) %>% 
  unique() %>% 
  mutate(genus = Genus, family = Family) %>% 
  dplyr::select(species, genus, family)
tree <- phylo.maker(sp.list = splist, tree = GBOTB.extended, nodes = nodes.info.1, scenarios = "S3")
ape::write.tree(tree$scenario.3, "save/phylogeny_full.tree")
```

```{r parsanb, eval=F}
load("save/growth.Rdata")  
pars <- as.data.frame(fit_full, pars = c('gmax')) %>% 
  reshape2::melt(variable.name = "parameter") %>% 
  group_by(parameter) %>% 
  summarise(value = median(value)) %>% 
  separate(parameter, c("parameter", "ind"), convert = T) %>% 
  reshape2::dcast(ind ~ parameter)
rm(fit_full) ; invisible(gc())
mdata <- vroom::vroom("save/mdatafull.tsv")
growth <- pars %>% 
  left_join(dplyr::select(mdata, ind, idTree, Family, Genus, species)) %>% 
  left_join(vroom::vroom(file = "save/env_full.tsv")) %>% 
  unique()
vroom::vroom_write(growth, file = "save/growthfull.tsv") 
```

# Analyses full

In this chapter, I investigated effects of phylogeny and ecological processes on individual growth, using phylogeny, topography and neighbourhood indices.

**Current results are using full data from chapter 4.**

## Methods

### Environmental descriptors

I used the mean neighbourhood crowding index [$NCI$; @Uriarte2004] over the last 30 years,
an indirect measurement of access to light and forest gap dynamics.
The mean neighbourhood crowding index $NCI_i$ from tree individual $i$ was calculated as follows:

$$NCI_i=\overline{\sum_{j|\delta_{i,j}<20m}DBH^2_{j,t}.e^{-\frac14\delta_{i,j}}}$$

with $DBH_{j,t}$ the diameter of the neighbouring tree $j$ in year $t$ and $\delta_{i,j}$ its distance to the individual tree $i$.
$NCI_i$ is computed for all neighbours at a distance $\delta_{i,j}$ inferior to the maximum neighbouring distance of 20 meters.
The power of neighbours $DBH_{j,t}$ effect was set to 2 to represent a surface.
The decrease of neighbours' diameter effect with distance was set to -0.25 to represent trees at 20 meters of the focal trees having 1% of the effect of the same tree at 0 meters.
$NCI_i$ is computed as the mean of yearly $NCI_{i,t}$ over the last 30 years denoted by the overline.

I used the topographic wetness index ($TWI$) as proxies of the distribution of soil water and nutrients in Paracou.
Waterlogging and topography have been highlighted as crucial for forest dynamics [@Ferry2010],
species-habitat relationships [@Engelbrecht2007], and phenotypic variation [@Schmitt2020].
Topography, through the dissolution of iron oxides, litter- and tree-fall transfers and waterlogging,
shapes soil nutrient distribution in tropical forests [@John2007; @Ferry2010].
TWI was derived from a 1-m-resolution digital elevation model using SAGA-GIS [@Conrad2015]
based on a LiDAR campaign of the whole Paracou field station done in 2015.

### Analyses

To study the effect of phylogeny and environment,
we investigated the effects of family, genus, species & topography $TWI$ and neighbourhood $NCI$ indices
on individual growth potential $Gmax$ with the following linear mixed model:

$$log(Gmax) ~ \sim \mathcal N (\alpha_{species} + \beta_{NCI} \times log(NCI),\sigma) \\ |~  \alpha_{species} \sim \mathcal N(\alpha_{genus},\sigma_{species})\\ |~  \alpha_{genus} \sim \mathcal N(\alpha_{family},\sigma_{genus})\\ |~  \alpha_{family} \sim \mathcal N(\alpha,\sigma_{family})$$

We reported the corresponding marginal and conditional $R^2$ [@Nakagawa2013].
We further plotted individual growth potential $Gmax$ along environmental gradients and across the phylogeny.
We used `phylosingal` to test for phylogenetic signal [@keck_phylosignal:_2016],
before computing phylogenetic correlogram and local indicator of phylogenetic association.

We used R version 3.6 for all statistical analyses (<www.r-project.org>). 

## Results

### General

Most of the variation in growth potential is individual ($\sigma_i = 0.437$), then explained by genus ($\sigma_g = 0.09$)
before species ($\sigma_s=0.04$) and relatively little by family ($\sigma_f=0.02$). 
The taxonomic structure explains almost a third of the observed variation in individual growth potential ($R^2_c=0.322$).
Finally, the neighbourhood ($NCI$) has a marked negative significant effect ($\beta=-0.49, p<0.001$, Fig. \@ref(fig:gmaxncib)) which explains 3% of the observed variation ($R^2_m=0.033$).

```{r dataanb}
growth <- vroom::vroom("save/growthfull.tsv")
phylo <- ape::read.tree("save/phylogeny_full.tree")
p4d <- phylo4d(phylo,
               data.frame(species = gsub("_", " ", phylo$tip.label)) %>% 
                 left_join(group_by(growth, species) %>% 
                             summarise(igv  = sqrt(exp(var(log(gmax)))-1), gmax  = median(gmax), logigv = log(igv), loggmax = log(gmax))
                           ) %>% 
                 dplyr::select(-species))   
```

```{r lmmb}
lme4::lmer(log(gmax) ~ log(NCI) + log(twi+1) + (1  | Family/Genus/species), growth) %>% 
  sjPlot::tab_model(show.icc = F)  
```

```{r r2b, eval=F }
m <- lme4::lmer(log(gmax) ~  log(NCI) + log(twi+1) + (1  | Family/Genus/species),
                na.omit(dplyr::select(growth, Family, Genus, species, NCI, twi, gmax)) %>% mutate(twi = ifelse(twi <= 0, 10^-6, twi)))
r2 <- partR2::partR2(m, partvars = c("log(NCI)", "log(twi+1)"), R2_type = "marginal", nboot = 10)
partR2::forestplot(r2, type = "R2", text_size = 10)
```

```{r, fig.width=4}
(g.var <- nlme::lme(log(gmax) ~  log(NCI) + log(twi+1), random=~1|Family/Genus/species,
          na.omit(dplyr::select(growth, Family, Genus, species, NCI, twi, gmax)) %>% mutate(twi = ifelse(twi <= 0, 10^-6, twi))) %>% 
  ape::varcomp(scale = F, cum = F) %>% 
  as.vector() %>% 
  data.frame(level = c("Family", "Genus", "Species", "Individual"), variance = as.vector(.)) %>% 
  select(-`.`) %>% 
  mutate(pct = variance / sum(variance)*100) %>% 
  mutate(level = factor(level, levels = c("Family", "Genus", "Species", "Individual"))) %>% 
  mutate(pct_text = paste0(round(pct), "%")) %>% 
  mutate(pct_text = gsub("^0%", "", pct_text)) %>% 
  ggplot(aes(x = 1, y = variance, fill = level)) +
  geom_bar(stat = "identity", position = "stack") +
  geom_text(aes(y = variance, label = pct_text), col = "white", position = position_stack(vjust = .5)) +
  scale_fill_discrete(expression(sigma^2)) +
  theme(axis.title = element_blank(), axis.line = element_blank(), axis.text = element_blank(), axis.ticks = element_blank())+
  ggtitle("B."))  
```


```{r gmaxncib, fig.cap="Relation between Neighbourhood crowding index (NCI) and individual growth potential (Gmax, cm/yr)."}
# MASS::rlm(log(gmax) ~ log(NCI), method="MM", growth) %>%
#   sjPlot::tab_model(show.icc = F)
# lm(log(gmax) ~ log(NCI), method="MM", growth) %>%
#   sjPlot::tab_model(show.icc = F)
(g.nci <-ggplot(growth, aes(NCI, gmax, col = species, group = NA)) +
  geom_bin2d(bins = 100) +
  scale_fill_continuous(type = "viridis") +
  geom_smooth(method = MASS::rlm, aes(group = NA), se = T, col = "red") +
  viridis::scale_fill_viridis(guide = "none") +
  scale_x_log10() + scale_y_log10() +
  xlab("Neighbourhood crowding index (NCI)") +
  ylab(expression(Gmax[i]~cm.year^-1)) +
  ggtitle("A.", expression(beta==-0.99~p<0.001~R^2==0.102))) 
```

### Phylogenetic

Species growth potential in the phylogeny (Fig. \@ref(fig:gmaxphylob)) is significantly structured (Tab. \@ref(tab:gmaxphylosignalb)),
with a short distance significant positive association and a long distant significant negative association (Fig. \@ref(fig:gmaxcorrelogb)).
Local indicator of phylogenetic association highlight the conservation of species growth potential at the genus level (Fig. \@ref(fig:lipab)),
as illustrated for instance with fast growing *Cecropia* opposed to slow growing *Eschweilera*.
However, a few species have different growth potential in the same genus, 
such as slow growing *Drypetes variabilis* opposed to fast growing *Drypetes fanshawei*.

```{r rlmigv}
MASS::rlm(cvlog_gmaxi ~ median_gmaxi, method="MM",
          summarise(group_by(growth, species), cvlog_gmaxi  = sqrt(exp(var(log(gmax)))-1), median_gmaxi  = median(gmax))) %>% 
  sjPlot::tab_model(show.icc = F)  
# lm(cvlog_gmaxi ~ median_gmaxi, method="MM",
#           summarise(group_by(growth, species), cvlog_gmaxi  = sqrt(exp(var(log(gmax)))-1), median_gmaxi  = median(gmax))) %>%
#   sjPlot::tab_model(show.icc = F)
```

```{r igvgmaxfig, fig.cap="Relation between species variance in growth potential (Gmax, cm/yr) and species median growth potential (Gmax, cm/yr)."}
(g.igv <- summarise(group_by(growth, species), cvlog_gmaxi  = sqrt(exp(var(log(gmax)))-1), median_gmaxi  = median(gmax),
          cvlog_nci  = sqrt(exp(var(log(NCI)))-1), median_nci  = median(NCI)) %>% 
  separate(species, c("Genus", "Species"), remove = F) %>% 
  ggplot(aes(median_gmaxi, cvlog_gmaxi)) +
  geom_point() +
  stat_smooth(method=MASS::rlm) +
  xlab(expression(Median(Gmax[i])~cm.year^-1)) + ylab(expression(CV[log](Gmax[i])~cm.year^-1)) +
  ggrepel::geom_text_repel(aes(label = paste0(str_sub(Genus, 1, 1), ". ", Species), alpha = median_gmaxi > 2)) +
  scale_alpha_manual(values = c(0, 0.8), guide = "none")  +
  ggtitle("X.", expression(beta==-0.07~p==0.018~R^2==0.015))) 
```

```{r gmaxphylob, fig.cap="Distribution of species growth potential (Gmax, cm/yr) in the phylogeny."}
(g.phylo <- fortify(phylo) %>% 
  mutate(species = gsub("_", " ", label)) %>%
  mutate(label = species) %>% 
  left_join(group_by(growth, species) %>% summarise(gmax = median(gmax))) %>% 
  ggtree(aes(color = gmax), layout="circular") + 
  geom_tiplab2(size = 2) +
  theme_tree(legend.position='right', legend.text = element_text(face = "italic")) +
  scale_alpha_manual("taxon", values = c(0.2, 1)) +
  scale_size_manual("taxon", values = c(1, 2)) +
  viridis::scale_color_viridis(expression(Median(Gmax[i])~cm.year^-1), trans = "log", labels=scales::comma) +
  theme(legend.position = "bottom") + xlim(NA, 200) +
  ggtitle("C.", expression(Pagel~lambda==0.72~p<0.001))) 
```

```{r igvphylob, fig.cap="Distribution of intraspecifig growth variability (IGV, cm/yr) in the phylogeny."}
fortify(phylo) %>% 
  mutate(species = gsub("_", " ", label)) %>%
  mutate(label = species) %>% 
  left_join(group_by(growth, species) %>% summarise(igv = sqrt(exp(var(log(gmax)))-1))) %>% 
  ggtree(aes(color = igv), layout="circular") + 
  geom_tiplab2(size = 2) +
  theme_tree(legend.position='right', legend.text = element_text(face = "italic")) +
  scale_alpha_manual("taxon", values = c(0.2, 1)) +
  scale_size_manual("taxon", values = c(1, 2)) +
  viridis::scale_color_viridis(expression(CV[log](Gmax[i])~cm.year^-1), trans = "log") + xlim(NA, 200)  
```

```{r gmaxphylosignalb}
phyloSignal(p4d = p4d, method = "all") %>% 
  lapply(as.data.frame) %>% 
  lapply(rownames_to_column, "parameter") %>% 
  bind_rows(.id = "type") %>% 
  reshape2::melt(c("type", "parameter")) %>% 
  reshape2::dcast(parameter + variable ~ type) %>% 
  mutate(value = paste0(round(stat, 4), " (p=", round(pvalue, 4), ")")) %>% 
  reshape2::dcast(parameter ~ variable) %>% 
  kable(caption = "Phylogenetic signal of species growth potential (Gmax, cm/yr) with different methods.")    
```

```{r gmaxcorrelogb, fig.cap="phylogenetic correlogram of species growth potential (Gmax, cm/yr)."}
# crlg <- phyloCorrelogram(p4d, trait = "loggmax", ci.bs	= 100)
# png(file = "save/figs/gmaxcorrelog_full.png", width = 1000, height = 1000, units = "px")
# plot(crlg, main="log(gmax) phylogenetic correlogram")
# dev.off()
include_graphics("save/figs/gmaxcorrelog_full.png") 
```

```{r igvcorrelogb, eval=T}
# crlg <- phyloCorrelogram(p4d, trait = "logigv", ci.bs	= 100)
# png(file = "save/figs/igvcorrelog_full.png", width = 1000, height = 1000, units = "px")
# plot(crlg, main="log(IGV) phylogenetic correlogram")
# dev.off()
include_graphics("save/figs/igvcorrelog_full.png") 
```

```{r lipab, fig.cap="Local phylogenetic associations of of species growth potential (Gmax, cm/yr) among taxa in the phylogeny."}
fortify(phylo) %>% 
  mutate(species = gsub("_", " ", label)) %>%
  mutate(label = species) %>% 
  left_join(lipaMoran(p4d)$p.value %>% 
              as.data.frame() %>% 
              rownames_to_column("species") %>% 
              mutate(species = gsub("_", " ", species)) %>% 
              dplyr::rename(pval = loggmax)) %>% 
  ggtree(aes(color = pval < 0.5), layout="circular") + 
  geom_tiplab2(size = 2) +
  theme_tree(legend.position='right', legend.text = element_text(face = "italic")) +
  scale_alpha_manual("taxon", values = c(0.2, 1)) +
  scale_size_manual("taxon", values = c(1, 2)) +
  scale_color_manual("Local\nIndicator\nof\nPhylogenetic\nAssociation\np<0.05", values =  c("grey", "red", "grey")) +
  xlim(NA, 200)   
```

```{r figall, fig.width=10, fig.height=7}
gridExtra::grid.arrange(
  grobs = list(g.phylo,  g.nci, g.var),
  layout_matrix = rbind(c(2, 2, 1, 1, 1),
                        c(3, 1, 1, 1, 1)),
  widths = c(1,0.5,1, 1, 1)
) 
```

```{r figall2, fig.width=10, fig.height=5}
gridExtra::grid.arrange(
  grobs = list(g.phylo,  g.nci),
  layout_matrix = rbind(c(2, 1)),
) 
```
